+++ mllib/src/main/scala/org/apache/spark/ml/recommendation/ALS.scala
@@ -23,420 +23,28 @@ import java.util.Locale
 
 import scala.collection.mutable
 import scala.reflect.ClassTag
-import scala.util.{Sorting, Try}
+import scala.util.Sorting
 import scala.util.hashing.byteswap64
 
 import com.github.fommil.netlib.BLAS.{getInstance => blas}
 import org.apache.hadoop.fs.Path
-import org.json4s.DefaultFormats
-import org.json4s.JsonDSL._
 
 import org.apache.spark.{Dependency, Partitioner, ShuffleDependency, SparkContext}
 import org.apache.spark.annotation.{DeveloperApi, Since}
 import org.apache.spark.internal.Logging
-import org.apache.spark.ml.{Estimator, Model}
-import org.apache.spark.ml.param._
-import org.apache.spark.ml.param.shared._
+import org.apache.spark.ml.Estimator
+import org.apache.spark.ml.param.ParamMap
 import org.apache.spark.ml.util._
 import org.apache.spark.mllib.linalg.CholeskyDecomposition
 import org.apache.spark.mllib.optimization.NNLS
 import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.{DataFrame, Dataset}
+import org.apache.spark.sql.Dataset
 import org.apache.spark.sql.functions._
-import org.apache.spark.sql.types._
+import org.apache.spark.sql.types.{FloatType, StructType}
 import org.apache.spark.storage.StorageLevel
 import org.apache.spark.util.Utils
 import org.apache.spark.util.collection.{OpenHashMap, OpenHashSet, SortDataFormat, Sorter}
 import org.apache.spark.util.random.XORShiftRandom
-
-/**
- * Common params for ALS and ALSModel.
- */
-private[recommendation] trait ALSModelParams extends Params with HasPredictionCol {
-  /**
-   * Param for the column name for user ids. Ids must be integers. Other
-   * numeric types are supported for this column, but will be cast to integers as long as they
-   * fall within the integer value range.
-   * Default: "user"
-   * @group param
-   */
-  val userCol = new Param[String](this, "userCol", "column name for user ids. Ids must be within " +
-    "the integer value range.")
-
-  /** @group getParam */
-  def getUserCol: String = $(userCol)
-
-  /**
-   * Param for the column name for item ids. Ids must be integers. Other
-   * numeric types are supported for this column, but will be cast to integers as long as they
-   * fall within the integer value range.
-   * Default: "item"
-   * @group param
-   */
-  val itemCol = new Param[String](this, "itemCol", "column name for item ids. Ids must be within " +
-    "the integer value range.")
-
-  /** @group getParam */
-  def getItemCol: String = $(itemCol)
-
-  /**
-   * Attempts to safely cast a user/item id to an Int. Throws an exception if the value is
-   * out of integer range or contains a fractional part.
-   */
-  protected[recommendation] val checkedCast = udf { (n: Any) =>
-    n match {
-      case v: Int => v // Avoid unnecessary casting
-      case v: Number =>
-        val intV = v.intValue
-        // Checks if number within Int range and has no fractional part.
-        if (v.doubleValue == intV) {
-          intV
-        } else {
-          throw new IllegalArgumentException(s"ALS only supports values in Integer range " +
-            s"and without fractional part for columns ${$(userCol)} and ${$(itemCol)}. " +
-            s"Value $n was either out of Integer range or contained a fractional part that " +
-            s"could not be converted.")
-        }
-      case _ => throw new IllegalArgumentException(s"ALS only supports values in Integer range " +
-        s"for columns ${$(userCol)} and ${$(itemCol)}. Value $n was not numeric.")
-    }
-  }
-
-  /**
-   * Param for strategy for dealing with unknown or new users/items at prediction time.
-   * This may be useful in cross-validation or production scenarios, for handling user/item ids
-   * the model has not seen in the training data.
-   * Supported values:
-   * - "nan":  predicted value for unknown ids will be NaN.
-   * - "drop": rows in the input DataFrame containing unknown ids will be dropped from
-   *           the output DataFrame containing predictions.
-   * Default: "nan".
-   * @group expertParam
-   */
-  val coldStartStrategy = new Param[String](this, "coldStartStrategy",
-    "strategy for dealing with unknown or new users/items at prediction time. This may be " +
-    "useful in cross-validation or production scenarios, for handling user/item ids the model " +
-    "has not seen in the training data. Supported values: " +
-    s"${ALSModel.supportedColdStartStrategies.mkString(",")}.",
-    (s: String) =>
-      ALSModel.supportedColdStartStrategies.contains(s.toLowerCase(Locale.ROOT)))
-
-  /** @group expertGetParam */
-  def getColdStartStrategy: String = $(coldStartStrategy).toLowerCase(Locale.ROOT)
-}
-
-/**
- * Common params for ALS.
- */
-private[recommendation] trait ALSParams extends ALSModelParams with HasMaxIter with HasRegParam
-  with HasPredictionCol with HasCheckpointInterval with HasSeed {
-
-  /**
-   * Param for rank of the matrix factorization (positive).
-   * Default: 10
-   * @group param
-   */
-  val rank = new IntParam(this, "rank", "rank of the factorization", ParamValidators.gtEq(1))
-
-  /** @group getParam */
-  def getRank: Int = $(rank)
-
-  /**
-   * Param for number of user blocks (positive).
-   * Default: 10
-   * @group param
-   */
-  val numUserBlocks = new IntParam(this, "numUserBlocks", "number of user blocks",
-    ParamValidators.gtEq(1))
-
-  /** @group getParam */
-  def getNumUserBlocks: Int = $(numUserBlocks)
-
-  /**
-   * Param for number of item blocks (positive).
-   * Default: 10
-   * @group param
-   */
-  val numItemBlocks = new IntParam(this, "numItemBlocks", "number of item blocks",
-      ParamValidators.gtEq(1))
-
-  /** @group getParam */
-  def getNumItemBlocks: Int = $(numItemBlocks)
-
-  /**
-   * Param to decide whether to use implicit preference.
-   * Default: false
-   * @group param
-   */
-  val implicitPrefs = new BooleanParam(this, "implicitPrefs", "whether to use implicit preference")
-
-  /** @group getParam */
-  def getImplicitPrefs: Boolean = $(implicitPrefs)
-
-  /**
-   * Param for the alpha parameter in the implicit preference formulation (nonnegative).
-   * Default: 1.0
-   * @group param
-   */
-  val alpha = new DoubleParam(this, "alpha", "alpha for implicit preference",
-    ParamValidators.gtEq(0))
-
-  /** @group getParam */
-  def getAlpha: Double = $(alpha)
-
-  /**
-   * Param for the column name for ratings.
-   * Default: "rating"
-   * @group param
-   */
-  val ratingCol = new Param[String](this, "ratingCol", "column name for ratings")
-
-  /** @group getParam */
-  def getRatingCol: String = $(ratingCol)
-
-  /**
-   * Param for whether to apply nonnegativity constraints.
-   * Default: false
-   * @group param
-   */
-  val nonnegative = new BooleanParam(
-    this, "nonnegative", "whether to use nonnegative constraint for least squares")
-
-  /** @group getParam */
-  def getNonnegative: Boolean = $(nonnegative)
-
-  /**
-   * Param for StorageLevel for intermediate datasets. Pass in a string representation of
-   * `StorageLevel`. Cannot be "NONE".
-   * Default: "MEMORY_AND_DISK".
-   *
-   * @group expertParam
-   */
-  val intermediateStorageLevel = new Param[String](this, "intermediateStorageLevel",
-    "StorageLevel for intermediate datasets. Cannot be 'NONE'.",
-    (s: String) => Try(StorageLevel.fromString(s)).isSuccess && s != "NONE")
-
-  /** @group expertGetParam */
-  def getIntermediateStorageLevel: String = $(intermediateStorageLevel)
-
-  /**
-   * Param for StorageLevel for ALS model factors. Pass in a string representation of
-   * `StorageLevel`.
-   * Default: "MEMORY_AND_DISK".
-   *
-   * @group expertParam
-   */
-  val finalStorageLevel = new Param[String](this, "finalStorageLevel",
-    "StorageLevel for ALS model factors.",
-    (s: String) => Try(StorageLevel.fromString(s)).isSuccess)
-
-  /** @group expertGetParam */
-  def getFinalStorageLevel: String = $(finalStorageLevel)
-
-  setDefault(rank -> 10, maxIter -> 10, regParam -> 0.1, numUserBlocks -> 10, numItemBlocks -> 10,
-    implicitPrefs -> false, alpha -> 1.0, userCol -> "user", itemCol -> "item",
-    ratingCol -> "rating", nonnegative -> false, checkpointInterval -> 10,
-    intermediateStorageLevel -> "MEMORY_AND_DISK", finalStorageLevel -> "MEMORY_AND_DISK",
-    coldStartStrategy -> "nan")
-
-  /**
-   * Validates and transforms the input schema.
-   *
-   * @param schema input schema
-   * @return output schema
-   */
-  protected def validateAndTransformSchema(schema: StructType): StructType = {
-    // user and item will be cast to Int
-    SchemaUtils.checkNumericType(schema, $(userCol))
-    SchemaUtils.checkNumericType(schema, $(itemCol))
-    // rating will be cast to Float
-    SchemaUtils.checkNumericType(schema, $(ratingCol))
-    SchemaUtils.appendColumn(schema, $(predictionCol), FloatType)
-  }
-}
-
-/**
- * Model fitted by ALS.
- *
- * @param rank rank of the matrix factorization model
- * @param userFactors a DataFrame that stores user factors in two columns: `id` and `features`
- * @param itemFactors a DataFrame that stores item factors in two columns: `id` and `features`
- */
-@Since("1.3.0")
-class ALSModel private[ml] (
-    @Since("1.4.0") override val uid: String,
-    @Since("1.4.0") val rank: Int,
-    @transient val userFactors: DataFrame,
-    @transient val itemFactors: DataFrame)
-  extends Model[ALSModel] with ALSModelParams with MLWritable {
-
-  /** @group setParam */
-  @Since("1.4.0")
-  def setUserCol(value: String): this.type = set(userCol, value)
-
-  /** @group setParam */
-  @Since("1.4.0")
-  def setItemCol(value: String): this.type = set(itemCol, value)
-
-  /** @group setParam */
-  @Since("1.3.0")
-  def setPredictionCol(value: String): this.type = set(predictionCol, value)
-
-  /** @group expertSetParam */
-  @Since("2.2.0")
-  def setColdStartStrategy(value: String): this.type = set(coldStartStrategy, value)
-
-  private val predict = udf { (featuresA: Seq[Float], featuresB: Seq[Float]) =>
-    if (featuresA != null && featuresB != null) {
-      // TODO(SPARK-19759): try dot-producting on Seqs or another non-converted type for
-      // potential optimization.
-      blas.sdot(rank, featuresA.toArray, 1, featuresB.toArray, 1)
-    } else {
-      Float.NaN
-    }
-  }
-
-  @Since("2.0.0")
-  override def transform(dataset: Dataset[_]): DataFrame = {
-    transformSchema(dataset.schema)
-    // create a new column named map(predictionCol) by running the predict UDF.
-    val predictions = dataset
-      .join(userFactors,
-        checkedCast(dataset($(userCol))) === userFactors("id"), "left")
-      .join(itemFactors,
-        checkedCast(dataset($(itemCol))) === itemFactors("id"), "left")
-      .select(dataset("*"),
-        predict(userFactors("features"), itemFactors("features")).as($(predictionCol)))
-    getColdStartStrategy match {
-      case ALSModel.Drop =>
-        predictions.na.drop("all", Seq($(predictionCol)))
-      case ALSModel.NaN =>
-        predictions
-    }
-  }
-
-  @Since("1.3.0")
-  override def transformSchema(schema: StructType): StructType = {
-    // user and item will be cast to Int
-    SchemaUtils.checkNumericType(schema, $(userCol))
-    SchemaUtils.checkNumericType(schema, $(itemCol))
-    SchemaUtils.appendColumn(schema, $(predictionCol), FloatType)
-  }
-
-  @Since("1.5.0")
-  override def copy(extra: ParamMap): ALSModel = {
-    val copied = new ALSModel(uid, rank, userFactors, itemFactors)
-    copyValues(copied, extra).setParent(parent)
-  }
-
-  @Since("1.6.0")
-  override def write: MLWriter = new ALSModel.ALSModelWriter(this)
-
-  /**
-   * Returns top `numItems` items recommended for each user, for all users.
-   * @param numItems max number of recommendations for each user
-   * @return a DataFrame of (userCol: Int, recommendations), where recommendations are
-   *         stored as an array of (itemCol: Int, rating: Float) Rows.
-   */
-  @Since("2.2.0")
-  def recommendForAllUsers(numItems: Int): DataFrame = {
-    recommendForAll(userFactors, itemFactors, $(userCol), $(itemCol), numItems)
-  }
-
-  /**
-   * Returns top `numUsers` users recommended for each item, for all items.
-   * @param numUsers max number of recommendations for each item
-   * @return a DataFrame of (itemCol: Int, recommendations), where recommendations are
-   *         stored as an array of (userCol: Int, rating: Float) Rows.
-   */
-  @Since("2.2.0")
-  def recommendForAllItems(numUsers: Int): DataFrame = {
-    recommendForAll(itemFactors, userFactors, $(itemCol), $(userCol), numUsers)
-  }
-
-  /**
-   * Makes recommendations for all users (or items).
-   * @param srcFactors src factors for which to generate recommendations
-   * @param dstFactors dst factors used to make recommendations
-   * @param srcOutputColumn name of the column for the source ID in the output DataFrame
-   * @param dstOutputColumn name of the column for the destination ID in the output DataFrame
-   * @param num max number of recommendations for each record
-   * @return a DataFrame of (srcOutputColumn: Int, recommendations), where recommendations are
-   *         stored as an array of (dstOutputColumn: Int, rating: Float) Rows.
-   */
-  private def recommendForAll(
-      srcFactors: DataFrame,
-      dstFactors: DataFrame,
-      srcOutputColumn: String,
-      dstOutputColumn: String,
-      num: Int): DataFrame = {
-    import srcFactors.sparkSession.implicits._
-
-    val ratings = srcFactors.crossJoin(dstFactors)
-      .select(
-        srcFactors("id"),
-        dstFactors("id"),
-        predict(srcFactors("features"), dstFactors("features")))
-    // We'll force the IDs to be Int. Unfortunately this converts IDs to Int in the output.
-    val topKAggregator = new TopByKeyAggregator[Int, Int, Float](num, Ordering.by(_._2))
-    val recs = ratings.as[(Int, Int, Float)].groupByKey(_._1).agg(topKAggregator.toColumn)
-      .toDF("id", "recommendations")
-
-    val arrayType = ArrayType(
-      new StructType()
-        .add(dstOutputColumn, IntegerType)
-        .add("rating", FloatType)
-    )
-    recs.select($"id" as srcOutputColumn, $"recommendations" cast arrayType)
-  }
-}
-
-@Since("1.6.0")
-object ALSModel extends MLReadable[ALSModel] {
-
-  private val NaN = "nan"
-  private val Drop = "drop"
-  private[recommendation] final val supportedColdStartStrategies = Array(NaN, Drop)
-
-  @Since("1.6.0")
-  override def read: MLReader[ALSModel] = new ALSModelReader
-
-  @Since("1.6.0")
-  override def load(path: String): ALSModel = super.load(path)
-
-  private[ALSModel] class ALSModelWriter(instance: ALSModel) extends MLWriter {
-
-    override protected def saveImpl(path: String): Unit = {
-      val extraMetadata = "rank" -> instance.rank
-      DefaultParamsWriter.saveMetadata(instance, path, sc, Some(extraMetadata))
-      val userPath = new Path(path, "userFactors").toString
-      instance.userFactors.write.format("parquet").save(userPath)
-      val itemPath = new Path(path, "itemFactors").toString
-      instance.itemFactors.write.format("parquet").save(itemPath)
-    }
-  }
-
-  private class ALSModelReader extends MLReader[ALSModel] {
-
-    /** Checked against metadata when loading model */
-    private val className = classOf[ALSModel].getName
-
-    override def load(path: String): ALSModel = {
-      val metadata = DefaultParamsReader.loadMetadata(path, sc, className)
-      implicit val format = DefaultFormats
-      val rank = (metadata.metadata \ "rank").extract[Int]
-      val userPath = new Path(path, "userFactors").toString
-      val userFactors = sparkSession.read.format("parquet").load(userPath)
-      val itemPath = new Path(path, "itemFactors").toString
-      val itemFactors = sparkSession.read.format("parquet").load(itemPath)
-
-      val model = new ALSModel(metadata.uid, rank, userFactors, itemFactors)
-
-      DefaultParamsReader.getAndSetParams(model, metadata)
-      model
-    }
-  }
-}
 
 /**
  * Alternating Least Squares (ALS) matrix factorization.
