+++ mllib/src/main/scala/org/apache/spark/ml/recommendation/ALSParams.scala
@@ -1,0 +1,156 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.ml.recommendation
+
+import scala.util.Try
+
+import org.apache.spark.ml.param._
+import org.apache.spark.ml.param.shared._
+import org.apache.spark.ml.util.SchemaUtils
+import org.apache.spark.sql.types.{FloatType, StructType}
+import org.apache.spark.storage.StorageLevel
+
+/**
+ * Common params for ALS.
+ */
+private[recommendation] trait ALSParams extends ALSModelParams with HasMaxIter with HasRegParam
+  with HasPredictionCol with HasCheckpointInterval with HasSeed {
+
+  /**
+   * Param for rank of the matrix factorization (positive).
+   * Default: 10
+   * @group param
+   */
+  val rank = new IntParam(this, "rank", "rank of the factorization", ParamValidators.gtEq(1))
+
+  /** @group getParam */
+  def getRank: Int = $(rank)
+
+  /**
+   * Param for number of user blocks (positive).
+   * Default: 10
+   * @group param
+   */
+  val numUserBlocks = new IntParam(this, "numUserBlocks", "number of user blocks",
+    ParamValidators.gtEq(1))
+
+  /** @group getParam */
+  def getNumUserBlocks: Int = $(numUserBlocks)
+
+  /**
+   * Param for number of item blocks (positive).
+   * Default: 10
+   * @group param
+   */
+  val numItemBlocks = new IntParam(this, "numItemBlocks", "number of item blocks",
+      ParamValidators.gtEq(1))
+
+  /** @group getParam */
+  def getNumItemBlocks: Int = $(numItemBlocks)
+
+  /**
+   * Param to decide whether to use implicit preference.
+   * Default: false
+   * @group param
+   */
+  val implicitPrefs = new BooleanParam(this, "implicitPrefs", "whether to use implicit preference")
+
+  /** @group getParam */
+  def getImplicitPrefs: Boolean = $(implicitPrefs)
+
+  /**
+   * Param for the alpha parameter in the implicit preference formulation (nonnegative).
+   * Default: 1.0
+   * @group param
+   */
+  val alpha = new DoubleParam(this, "alpha", "alpha for implicit preference",
+    ParamValidators.gtEq(0))
+
+  /** @group getParam */
+  def getAlpha: Double = $(alpha)
+
+  /**
+   * Param for the column name for ratings.
+   * Default: "rating"
+   * @group param
+   */
+  val ratingCol = new Param[String](this, "ratingCol", "column name for ratings")
+
+  /** @group getParam */
+  def getRatingCol: String = $(ratingCol)
+
+  /**
+   * Param for whether to apply nonnegativity constraints.
+   * Default: false
+   * @group param
+   */
+  val nonnegative = new BooleanParam(
+    this, "nonnegative", "whether to use nonnegative constraint for least squares")
+
+  /** @group getParam */
+  def getNonnegative: Boolean = $(nonnegative)
+
+  /**
+   * Param for StorageLevel for intermediate datasets. Pass in a string representation of
+   * `StorageLevel`. Cannot be "NONE".
+   * Default: "MEMORY_AND_DISK".
+   *
+   * @group expertParam
+   */
+  val intermediateStorageLevel = new Param[String](this, "intermediateStorageLevel",
+    "StorageLevel for intermediate datasets. Cannot be 'NONE'.",
+    (s: String) => Try(StorageLevel.fromString(s)).isSuccess && s != "NONE")
+
+  /** @group expertGetParam */
+  def getIntermediateStorageLevel: String = $(intermediateStorageLevel)
+
+  /**
+   * Param for StorageLevel for ALS model factors. Pass in a string representation of
+   * `StorageLevel`.
+   * Default: "MEMORY_AND_DISK".
+   *
+   * @group expertParam
+   */
+  val finalStorageLevel = new Param[String](this, "finalStorageLevel",
+    "StorageLevel for ALS model factors.",
+    (s: String) => Try(StorageLevel.fromString(s)).isSuccess)
+
+  /** @group expertGetParam */
+  def getFinalStorageLevel: String = $(finalStorageLevel)
+
+  setDefault(rank -> 10, maxIter -> 10, regParam -> 0.1, numUserBlocks -> 10, numItemBlocks -> 10,
+    implicitPrefs -> false, alpha -> 1.0, userCol -> "user", itemCol -> "item",
+    ratingCol -> "rating", nonnegative -> false, checkpointInterval -> 10,
+    intermediateStorageLevel -> "MEMORY_AND_DISK", finalStorageLevel -> "MEMORY_AND_DISK",
+    coldStartStrategy -> "nan")
+
+  /**
+   * Validates and transforms the input schema.
+   *
+   * @param schema input schema
+   * @return output schema
+   */
+  protected def validateAndTransformSchema(schema: StructType): StructType = {
+    // user and item will be cast to Int
+    SchemaUtils.checkNumericType(schema, $(userCol))
+    SchemaUtils.checkNumericType(schema, $(itemCol))
+    // rating will be cast to Float
+    SchemaUtils.checkNumericType(schema, $(ratingCol))
+    SchemaUtils.appendColumn(schema, $(predictionCol), FloatType)
+  }
+}
